refactor: mvc패턴으로 폴더구조 적용 및 불필요한 import문 제거
