+++ src/main/java/car/CarApplication.java
@@ -1,5 +1,8 @@
 package car;
 
+import car.controller.GameController;
+import car.model.RandomNumberGenerator;
+
 public class CarApplication {
     public static void main(String[] args) {
         GameController gameController = new GameController(new RandomNumberGenerator());
