+++ src/main/java/car/controller/InputParser.java
@@ -1,4 +1,4 @@
-package car;
+package car.controller;
 
 import java.util.List;
