+++ src/main/java/car/model/CarPosition.java
@@ -1,4 +1,4 @@
-package car;
+package car.model;
 
 import java.util.Objects;
