+++ src/main/java/car/model/RandomNumberGenerator.java
@@ -1,8 +1,6 @@
-package car;
+package car.model;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
