+++ src/main/java/car/view/InputView.java
@@ -1,14 +1,14 @@
-package car;
+package car.view;
 
 import java.util.Scanner;
 
 public class InputView {
     
-    public static Integer inputNumberOfCars() {
-        System.out.println("자동차 대수는 몇 대 인가요?");
-        Scanner scanner = new Scanner(System.in);
-        return scanner.nextInt();
-    }
+//    public static Integer inputNumberOfCars() {
+//        System.out.println("자동차 대수는 몇 대 인가요?");
+//        Scanner scanner = new Scanner(System.in);
+//        return scanner.nextInt();
+//    }
   
     public static Integer inputNumberOfGames() {
         System.out.println("시도할 회수는 몇 회 인가요?");
