+++ src/test/java/car/controller/InputParserTest.java
@@ -1,4 +1,4 @@
-package car;
+package car.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
