+++ src/test/java/car/model/CarPositionTest.java
@@ -1,4 +1,4 @@
-package car;
+package car.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
